backend/requirements.txt
fastapi
uvicorn[standard]
stripe
sqlalchemy
alembic
pydantic
python-dotenv
psycopg2
sqlite